- remove treeRef - remove unused tobeRenderedItemsProps null check - separate context - useTree takes UseTreeOptions
