+++ packages/fluentui/react-northstar/src/components/Tree/hooks/useTree.tsx
@@ -1,10 +1,31 @@
 import * as React from 'react';
-import { TreeProps } from '../Tree';
+import { ObjectShorthandCollection, ShorthandRenderFunction, TreeItemProps, TreeTitleProps } from '../../../index';
 import { flattenTree, getTobeRenderedItemsProps } from './flattenTree';
 import { useTreeActiveState } from './useTreeActiveState';
 import { useTreeBehavior } from './useTreeBehavior';
 
-export function useTree(props: TreeProps) {
+export interface UseTreeOptions {
+  /** Shorthand array of props for Tree. */
+  items?: ObjectShorthandCollection<TreeItemProps>;
+
+  /**
+   * A custom render function for the title slot.
+   *
+   * @param Component - The computed component for this slot.
+   * @param props - The computed props for this slot.
+   * @param children - The computed children for this slot.
+   */
+  renderItemTitle?: ShorthandRenderFunction<TreeTitleProps>;
+
+  /** Ids of expanded items. */
+  activeItemIds?: string[];
+  /** Initial activeItemIds value. */
+  defaultActiveItemIds?: string[];
+  /** Only allow one subtree to be expanded at a time. */
+  exclusive?: boolean;
+}
+
+export function useTree(props: UseTreeOptions) {
   // build flattened tree from props.items
   const flatTreeFromProps = React.useMemo(() => flattenTree(props.items), [props.items]);
